+++ newton/_src/wheeled/__init__.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 The Newton Developers
 # SPDX-License-Identifier: Apache-2.0
 
+from .contact_patch import WheelContactPatchState, update_wheel_contact_patches
 from .metadata import (
     WheeledAssetMetadata,
     WheeledModelMetadata,
@@ -14,6 +15,7 @@ from .metadata import (
 )
 
 __all__ = [
+    "WheelContactPatchState",
     "WheelMetadata",
     "WheeledAssetMetadata",
     "WheeledModelMetadata",
@@ -23,4 +25,5 @@ __all__ = [
     "load_wheeled_manifest",
     "read_wheeled_metadata",
     "register_wheeled_custom_attributes",
+    "update_wheel_contact_patches",
 ]